Fixes / Refactoring
