+++ extensions/Anthropic/Client/ContentBlockDelta.cs
@@ -1,0 +1,17 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+
+using System.Text.Json.Serialization;
+
+namespace Microsoft.KernelMemory.AI.Anthropic.Client;
+
+internal sealed class ContentBlockDelta : StreamingResponseMessage
+{
+    [JsonPropertyName("type")]
+    public string? Type { get; set; }
+
+    [JsonPropertyName("index")]
+    public int Index { get; set; }
+
+    [JsonPropertyName("delta")]
+    public Delta Delta { get; set; } = null!;
+}
