+++ extensions/Anthropic/Client/MessageRequest.cs
@@ -1,0 +1,26 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+
+using System.Text.Json.Serialization;
+
+namespace Microsoft.KernelMemory.AI.Anthropic.Client;
+
+internal sealed class MessageRequest
+{
+    [JsonPropertyName("model")]
+    public string? Model { get; set; }
+
+    [JsonPropertyName("stream")]
+    public bool Stream { get; set; }
+
+    [JsonPropertyName("max_tokens")]
+    public int MaxTokens { get; set; }
+
+    [JsonPropertyName("temperature")]
+    public double Temperature { get; set; }
+
+    [JsonPropertyName("system")]
+    public string? System { get; set; }
+
+    [JsonPropertyName("messages")]
+    public Message[]? Messages { get; set; }
+}
