+++ extensions/Anthropic/Client/StreamingResponseMessage.cs
@@ -1,0 +1,7 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+
+namespace Microsoft.KernelMemory.AI.Anthropic.Client;
+
+internal abstract class StreamingResponseMessage
+{
+}
